Create bike_class_with_performance.py
This Python program demonstrates the use of object-oriented programming concepts such as constructors (__init__) and instance methods. The Bikes class initializes bike attributes (model and colour) and includes a show_performance method to display key performance specifications of the Royal Enfield Hunter 350. This example illustrates how to store and retrieve object data as well as define object behavior in Python.